timetree-scheduler/backend/requirements.txt
# FastAPI and ASGI
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic
pydantic[email]==2.5.1
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Date/Time handling
python-dateutil==2.8.2
pendulum==2.1.2
pytz==2023.3

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0

# AI/ML
openai==1.54.0

# Redis (optional)
redis[hiredis]==5.0.1
aioredis==2.0.1

# Task Queue (optional)
celery[redis]==5.3.4

# Monitoring
sentry-sdk[fastapi]==1.38.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-otlp==1.21.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # for testing
factory-boy==3.3.0

# Code quality
ruff==0.1.6
black==23.11.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Performance
orjson==3.9.10